Refactor and add comments.
