+++ Cargo.toml
@@ -11,4 +11,3 @@ reqwest = "0.8.5"
 serde = "1.0.43"
 serde_derive = "1.0.43"
 serde_json = "1.0"
-rayon = "1.0.1"
