+++ src/hash.rs
@@ -11,15 +11,18 @@ impl ImageHash {
         ImageHash { bits }
     }
 
+    /// Get the Hamming distance between another ImageHash
     pub fn hamming(&self, other: &ImageHash) -> u32 {
         (self.bits ^ other.bits).count_ones()
     }
 
+    /// Get the similarity to another ImageHash as a float between 0 and 1, where 0 represents no similarity and 1 represents no difference.
     pub fn similarity(&self, other: &ImageHash) -> f32 {
-        1f32 - (self.hamming(other) as f32 / 64f32)
+        1f32 - (self.hamming(other) as f32 / RESIZE_DIMENSION.pow(2) as f32)
     }
 }
 
+/// Calculate the perceptual hash of an image using the Average Hash algorithm.
 pub fn average_hash(image: &DynamicImage) -> ImageHash {
     let resize = image
         .resize_exact(RESIZE_DIMENSION, RESIZE_DIMENSION, FilterType::Nearest)
@@ -27,9 +30,9 @@ pub fn average_hash(image: &DynamicImage) -> ImageHash {
 
     let raw = resize.raw_pixels();
 
-    assert_eq!(raw.len(), 64);
+    assert_eq!(raw.len(), RESIZE_DIMENSION.pow(2) as usize);
 
-    let average = raw.iter().map(|&n| n as usize).sum::<usize>() / 64;
+    let average = raw.iter().map(|&n| n as usize).sum::<usize>() / RESIZE_DIMENSION.pow(2) as usize;
 
     let hash = raw.iter()
         .map(|&n| (n as usize > average) as u64)
